+++ ryven/ironflow/__init__.py
@@ -1,4 +1,4 @@
-__all__ = ['gui', 'flow_canvas', 'canvas_widgets', 'node_widgets']
+__all__ = ['gui', 'flow_canvas', 'canvas_widgets', 'node_interface.py']
 
 from .gui import GUI
 from ryven.NENV import Node, NodeInputBP, NodeOutputBP, dtypes
